feat: v0.87 gensym/read/macroexpand + P2 基础层清理
v0.87 Lisp 系 homoiconicity（gensym/read/macroexpand）：
- CoreRuntime.gensym_counter Arc<Mutex<usize>> + 3 dispatch arm + 3 handler
-gensym/read/macroexpand 完整实现，19 个 v0.87 测试全过
- DAG volatile 用多 let 绑定规避

P2 基础层清理（2026-08-14→15）：
- 移除 AiConfigValue::budget/per_call 死字段（TokenBudget 独立管理）
- 移除 create_mock_stream 死函数（v0.04 占位，无调用方）
- 移除 extract_embeddings 死函数（v0.04 占位，无调用方）
- 移除 EnvRef::from_arc_mutex 遗留迁移方法
- 移除 RouteConfig::priority/healthy 死字段
- 移除 interpreter/ai_helpers.rs 中的 extract_embeddings
- 重命名 numeric_tower_number_float_compat → numeric_tower_float_plus_float_integer_values
- 清理 interpreter/builtins/ 下所有 #![allow(unused_mut)]（event/exec/memory/mod）
- cargo fix --lib 自动移除 10 个不必要的 mut
- 手动移除 2 个模块级 allow + 3 个 let mut → let interp（不改变语义）

验证：cargo clippy --all-targets --all-features -- -D warnings
     cargo test --lib（781 passed, 13 ignored）
     cargo test --test e2e（22 passed, 1 pre-existing tea_app 失败）

+++ src/typeck/dispatch.rs
@@ -80,16 +80,16 @@ pub fn builtin_signatures() -> Vec<(String, Signature)> {
                 raw_return_type: None,
             },
         ),
-        // range(start, end, step) -> list<number>
+        // range(start, end, step) -> list<int>
         (
             "range".to_string(),
             Signature::new(
                 vec![
-                    ("start".to_string(), Type::Float),
-                    ("end".to_string(), Type::Float),
-                    ("step".to_string(), Type::Float),
+                    ("start".to_string(), Type::Int),
+                    ("end".to_string(), Type::Int),
+                    ("step".to_string(), Type::Int),
                 ],
-                Type::List(Box::new(Type::Float)),
+                Type::List(Box::new(Type::Int)),
             ),
         ),
         // len(x) for string / list / dict
